Custom2LRUCache
This is a custom solution that uses a LinkedList for keeping the order
of the items, as required by LRU, and a HashMap for ensuring O(1) time
complexity for getting and putting a new item. Updating an existing item
will be O(n) as for bumping the existing item to the top of the
LinkedList requires a linear search.

It's important to note that this implementation assumes that there won't
be duplicate values in the cache, even if with different keys. Duplicate
values might result in a wrong item removed from the cache as there is
no way to determine the correct value in the LinkedList.

The unit test givenFullCacheWithTwoValuesUnderDifferentKeys_whenFirstItemRemoved_thenFirstItemShouldNotBeInTheCache
is falling due to this problem.