Custom3LRUCache
iThis is a custom solution that uses a LinkedList for keeping the order
of the items, as required by LRU, and a HashMap for ensuring O(1) time
complexity for getting and putting a new item. Updating an existing item
will be O(n) because bumping the existing item to the top of the
LinkedList requires a linear search.

It's important to note that this solution fixes the problem from the
previous by having a linked list of Nodes. All unit tests are now
passing.