Custom4LRUCache
This is a custom solution that uses a custom linked list for keeping the
order of the items, as required by LRU, and a HashMap for ensuring O(1)
time complexity for searching, getting and putting a new item. Updating
an existing item takes O(n) time complexity because bumping it to the
top of the linked list requires a linear search.

This solution differs from the previous by having a custom linked list,
rather using the Java's LinkedList. It proves my capacity of implementing
such data structure.